pyproject.toml
[project]
name = "switchbox-backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.40",
    "uvicorn[standard]>=0.41.0",
    "boto3>=1.35.0",
    "email-validator>=2.3.0",
    "pyjwt>=2.8.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.15.5",
    "pytest-cov>=6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app"]
